requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pandas
scikit-learn
openai
PyJWT==2.10.1
pwdlib[argon2]
python-multipart